Fix remaining warnings
