+++ .github/workflows/pull_request.yml
@@ -1,0 +1,17 @@
+#
+# This source file is part of the Stanford HealthGPT project
+#
+# SPDX-FileCopyrightText: 2023 Stanford University & Project Contributors
+#
+
+
+name: Pull Request
+
+on:
+  pull_request:
+  workflow_dispatch:
+
+jobs:
+  swiftlint:
+    name: SwiftLint
+    uses: StanfordBDHG/.github/.github/workflows/swiftlint.yml@v2
