+++ HealthGPT/HealthGPT/MessageInputView.swift
@@ -18,28 +18,37 @@ struct MessageInputView: View {
     @State private var alertText = ""
 
     private var apiKey: String? {
-        guard let filePath = Bundle.main.path(forResource: "OpenAI-Info", ofType: "plist") else {
+        guard let filePath = Bundle.main.path(forResource: "OpenAI-Info", ofType: "plist"),
+              let data = FileManager.default.contents(atPath: filePath) else {
             alertText = "Couldn't find file 'OpenAI-Info.plist'."
             self.showAlert.toggle()
             return nil
         }
 
-        let plist = NSDictionary(contentsOfFile: filePath)
-        guard let value = plist?.object(forKey: "API_KEY") as? String else {
-            alertText = "Couldn't find key 'API_KEY' in 'OpenAI-Info.plist'."
-            self.showAlert.toggle()
-            return nil
-        }
+        do {
+            guard let plist = try PropertyListSerialization.propertyList(from: data, options: [], format: nil) as? [String: Any],
+                  let value = plist["API_KEY"] as? String else {
+                alertText = "Couldn't find key 'API_KEY' in 'OpenAI-Info.plist'."
+                self.showAlert.toggle()
+                return nil
+            }
 
-        if value.starts(with: "_") {
-            alertText = "Please register for an OpenAI account and get an API Key. "
+            if value.starts(with: "_") {
+                alertText = "Please register for an OpenAI account and get an API Key. "
+                self.showAlert.toggle()
+                return nil
+            }
+            return value
+        } catch {
+            alertText = "An error occurred while reading the 'OpenAI-Info.plist' file."
             self.showAlert.toggle()
             return nil
         }
-        return value
     }
 
     var body: some View {
+        // This code is to be refactored, will temporarily disable.
+        // swiftlint:disable closure_body_length
         HStack {
             TextField(
                 isQuerying ? "HealthGPT is thinking 🤔..." : "Type a message...",
@@ -139,15 +148,31 @@ struct MessageInputView: View {
 
                             // Wait for both async calls to finish
                             group.notify(queue: .main) {
-                                
                                 var mainPrompt = "You are HealthGPT, an enthusiastic, expert caretaker with a deep understanding in personal health. Given the context, provide a short response that could answer the user's question. Do NOT provide statistics. If numbers seem low, provide advice on how they can improve.\n\nSome health metrics over the past two weeks (14 days) to incorporate is given below. If a value is zero, the user has not inputted anything for that day. Today is \(DateFormatter.localizedString(from: today, dateStyle: .full, timeStyle: .none)). Note that you do not have data about the current day.\n\n"
 
                                 for day in 0...13 {
                                     let dayData = healthData[day]
-                                    mainPrompt += "\(dayData.date): \(Int(dayData.steps!)) steps, \(Int(dayData.sleepHours!)) hours of sleep, \(Int(dayData.activeEnergy!) ) calories burned, \(Int(dayData.exerciseMinutes!) ) minutes of exercise, \(dayData.bodyWeight ?? 0.0) lbs of body weight\n"
+                                    var mainText = ""
+
+                                    if let steps = dayData.steps {
+                                        mainText += "\(Int(steps)) steps,"
+                                    }
+                                    if let sleepHours = dayData.sleepHours {
+                                        mainText += " \(Int(sleepHours)) hours of sleep,"
+                                    }
+                                    if let activeEnergy = dayData.activeEnergy {
+                                        mainText += " \(Int(activeEnergy)) calories burned,"
+                                    }
+                                    if let bodyWeight = dayData.bodyWeight {
+                                        mainText += " \(bodyWeight) lbs of body weight,"
+                                    }
+                                    if let exerciseMinutes = dayData.exerciseMinutes {
+                                        mainText += " \(Int(exerciseMinutes)) minutes of exercise,"
+                                    }
+
+                                    mainPrompt += "\(dayData.date):\(mainText.dropLast()) \n"
                                 }
                                 print(mainPrompt)
-
 
                                 Task {
                                     var currentChat: [Chat] = [.init(role: .system, content: mainPrompt)]
@@ -169,7 +194,6 @@ struct MessageInputView: View {
                                     isQuerying = false
                                 }
                             }
-
                         } else {
                             print("Authorization failed.")
                             isQuerying = false
@@ -203,30 +227,5 @@ struct MessageInputView: View {
                 dismissButton: .default(Text("OK"))
             )
         }
-    }
-}
-
-struct SettingsView: View {
-    @Environment(\.dismiss) var dismiss
-    @Binding var messages: [Message]
-
-    var body: some View {
-        Button("Clear Current Thread") {
-            messages = []
-            dismiss()
-        }
-        .padding()
-        .background(.white)
-        .cornerRadius(20)
-        .foregroundColor(.red)
-        .overlay(
-            RoundedRectangle(cornerRadius: 20)
-                .stroke(.red, lineWidth: 1)
-        )
-
-        Text("HealthGPT is powered by the OpenAI API. Data submitted here is not used for training OpenAI's models according to their terms and conditions.\n\nCurrently, HealthGPT is accessing your step count, sleep analysis, exercise minutes, active calories burned, body weight, and heart rate, all from data stored in the Health app.\n\nRemember to log your data and wear your Apple Watch throughout the day for the most accurate results.")
-            .foregroundColor(.gray)
-            .padding(20)
-            .font(.system(size: 15))
     }
 }
